+++ src/asp/Sessions/StereoSessionIp.cc
@@ -20,6 +20,7 @@
 
 /// \file StereoSessionIp.cc
 ///
+#include <asp/Sessions/StereoSessionASTER.h>
 #include <vw/Core/Exception.h>
 #include <vw/Core/Log.h>
 #include <vw/Math/Vector.h>
@@ -64,6 +65,20 @@ bool StereoSession::ip_matching(std::string const& input_file1,
                                 std::string const  right_ip_file) {
 
   vw_out() << "\t--> Matching interest points in StereoSession.\n";
+
+  // Fix for ASTER. This will modify the pointers to the cameras for this function,
+  // but not the cameras in the caller of this function.
+  boost::shared_ptr<camera::CameraModel> left_cam, right_cam;
+  if (this->name() == "aster") {
+    vw_out() << "Using the RPC model instead of the exact ASTER model for interest point "
+              << "matching, for speed. This does not affect the accuracy of final results.\n";
+    StereoSessionASTER * aster_session = dynamic_cast<StereoSessionASTER*>(this);
+    if (aster_session == NULL) 
+      vw_throw( ArgumentErr() << "ASTER session is expected." );
+    aster_session->rpc_camera_models(left_cam, right_cam);
+    cam1 = left_cam.get();
+    cam2 = right_cam.get();
+  }
 
   // Sanity checks. Must be here since we will use this code in stereo and bundle_adjust.
   if (asp::stereo_settings().matches_per_tile > 0) {